+++ client/vue/pages/HomePage.vue
@@ -32,7 +32,7 @@
 	</ul>
 	<div class="footer-container">
 		<div class="version" @click="showVersionHistory">
-			Marbleland v1.11.0
+			Marbleland v1.12.0
 		</div>
 		<router-link to="/support" class="support-button">Support</router-link>
 	</div>