+++ version_history.md
@@ -1,5 +1,9 @@
 ### Version history
 
+## 1.12.0
+- Added level updating and level versioning
+- Added "favorite levels" section to the profile page
+
 ## 1.11.1
 - Added support for filtering games and leaderboards based on exact datablock compatibility values as well as custom code.